+++ src/ts/errors.ts
@@ -1,7 +1,7 @@
-export { Err } from "./errors/BaseErr";
-export { PanicErr } from "./errors/PanicErr";
-export { UnknownErr } from "./errors/UnknownErr";
-export { IoErr } from "./errors/IoErr";
-export { InitErr } from "./errors/InitErr";
+export { Err } from "./errors/base_err";
+export { PanicErr } from "./errors/panic_err";
+export { UnknownErr } from "./errors/unknown_err";
+export { IoErr } from "./errors/io_err";
+export { InitErr } from "./errors/init_err";
 export { Result } from "./errors/Result";
 export { panic, wrapErr } from "./errors/utils";
