+++ src/ts/errors/base_err.ts
