+++ src/ts/errors/panic_err.ts
@@ -1,4 +1,4 @@
-import { Err } from "./BaseErr";
+import { Err } from "./base_err";
 
 export class PanicErr extends Err {
   constructor(message: string, cause?: Err) {
