+++ src/ts/errors/unknown_err.ts
@@ -1,4 +1,4 @@
-import { Err } from "./BaseErr";
+import { Err } from "./base_err";
 
 export class UnknownErr extends Err {
   constructor(cause?: Error) {
