+++ src/ts/rest/middleware/catch_unexpected_err.ts
@@ -1,6 +1,6 @@
 import { NextFunction } from "express";
 import { wrapErr } from "../../errors";
-import { ApiErr } from "../ApiErr";
+import { ApiErr } from "../api_err";
 import { Request } from "../Request";
 import { Response } from "../Response";
 import { status } from "../status";
