+++ src/ts/rest/middleware/send_response.ts
