+++ src/ts/rest/middleware/validate_request.ts
@@ -1,7 +1,7 @@
 import { NextFunction } from "express";
 import typing from "io-ts";
 import { validateType } from "../../typing";
-import { ApiErr } from "../ApiErr";
+import { ApiErr } from "../api_err";
 import { Request } from "../Request";
 import { Response } from "../Response";
 import { status } from "../status";
